requirements.txt
# Core agent framework
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pydantic>=2.0
python-dotenv>=1.0

# FastAPI backend
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9

# Data sources (all free)
yfinance>=0.2.40
requests>=2.31
beautifulsoup4>=4.12
PyGithub>=2.1
sec-edgar-downloader>=5.0

# Reporting
reportlab>=4.0
matplotlib>=3.8
pandas>=2.0

# Frontend
streamlit>=1.30

# Dev
pytest>=8.0
httpx>=0.27
